+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Moved Gallery view toggle button into header.
 - Removed css entries that are not needed anymore for Gallery View Toggle.
+- Change search box text when enabling advanced search.
 
 ### Advanced Search Implementation